refactor: 파일 폴더 구성 수정 등 리뷰반영
